requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
astroid==2.15.6
    # via pylint
attrs==23.1.0
    # via
    #   cattrs
    #   jsii
aws-assume-role-lib==2.10.0
    # via -r requirements.in
aws-cdk-asset-awscli-v1==2.2.200
    # via aws-cdk-lib
aws-cdk-asset-kubectl-v20==2.1.2
    # via aws-cdk-lib
aws-cdk-asset-node-proxy-agent-v5==2.0.166
    # via aws-cdk-lib
aws-cdk-lib==2.88.0
    # via -r requirements.in
aws-lambda-powertools==2.20.0
    # via -r requirements.in
boto3==1.28.8
    # via
    #   -r requirements.in
    #   aws-assume-role-lib
botocore==1.31.8
    # via
    #   boto3
    #   s3transfer
cattrs==23.1.2
    # via jsii
constructs==10.2.69
    # via
    #   -r requirements.in
    #   aws-cdk-lib
dill==0.3.6
    # via pylint
exceptiongroup==1.1.2
    # via
    #   cattrs
    #   pytest
importlib-resources==6.0.0
    # via jsii
iniconfig==2.0.0
    # via pytest
isort==5.12.0
    # via pylint
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsii==1.85.0
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v5
    #   aws-cdk-lib
    #   constructs
lazy-object-proxy==1.9.0
    # via astroid
mccabe==0.7.0
    # via pylint
packaging==23.1
    # via pytest
platformdirs==3.9.1
    # via pylint
pluggy==1.2.0
    # via pytest
publication==0.0.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v5
    #   aws-cdk-lib
    #   constructs
    #   jsii
pylint==2.17.4
    # via -r requirements.in
pylint-exit==1.2.0
    # via -r requirements.in
pytest==7.4.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   botocore
    #   jsii
s3transfer==0.6.1
    # via boto3
six==1.16.0
    # via python-dateutil
tomli==2.0.1
    # via
    #   pylint
    #   pytest
tomlkit==0.11.8
    # via pylint
typeguard==2.13.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v5
    #   aws-cdk-lib
    #   constructs
    #   jsii
typing-extensions==4.7.1
    # via
    #   astroid
    #   aws-lambda-powertools
    #   cattrs
    #   jsii
urllib3==1.26.16
    # via botocore
wrapt==1.15.0
    # via astroid
